Factor IsolateLoader out of Loader
This patch moves the Isolate-related concerns out of the main `Loader` class
and into an `IsolateLoader` class. This change makes the structure between
Isolate- and Browser-based loading more parallel and paves the way to introduce
a `SubprocessLoader`.
